requirements.txt
google-auth==2.48.0
google-genai==1.62.0
graphviz==0.21
Jinja2==3.1.6
jiter==0.13.0
numpy==2.2.6
openai==2.17.0
python-dotenv==1.2.1
PyYAML==6.0.3
streamlit==1.54.0
openpyxl==3.1.5